SMACK-458 Managers should be kept on disconnects
Smack's Managers should not remove itself when the connection is closed
or should re-add themselfs if the connection get reconnected.

This should also fix some NPE's.

We are currenlty going with two different designs of Manager: 1. The one
with WeakReferences (SDM, EntityCapsManager) and 2. the one where the
managers remove their listeners on connectionClosed() *and*
connectionClosedOnError(), and later add their listeners on
reconnectionSuccessful().